Adding a present common transitioning delegate method and content view controller
