+++ Client/Frontend/Browser/SendTab/SendTabToSelfController.swift
@@ -8,7 +8,7 @@ import BraveUI
 import BraveCore
 import BraveShared
 
-class SendTabToSelfController: UIViewController {
+class SendTabToSelfController: SendTabTransitioningController {
   
   struct UX {
     static let contentInset = 20.0
@@ -18,10 +18,6 @@ class SendTabToSelfController: UIViewController {
   let contentNavigationController: UINavigationController
   private let sendTabContentController: SendTabToSelfContentController
   
-  let backgroundView = UIView().then {
-    $0.backgroundColor = UIColor(white: 0.0, alpha: 0.3)
-  }
-
   init(sendTabAPI: BraveSendTabAPI, dataSource: SendableTabInfoDataSource) {
     sendTabContentController = SendTabToSelfContentController(sendTabAPI: sendTabAPI, dataSource: dataSource)
     contentNavigationController = UINavigationController(rootViewController: sendTabContentController).then {
@@ -30,10 +26,8 @@ class SendTabToSelfController: UIViewController {
       $0.view.clipsToBounds = true
     }
     
-    super.init(nibName: nil, bundle: nil)
+    super.init()
         
-    transitioningDelegate = self
-    modalPresentationStyle = .overFullScreen
     addChild(contentNavigationController)
     contentNavigationController.didMove(toParent: self)
   }
@@ -46,9 +40,7 @@ class SendTabToSelfController: UIViewController {
   override func viewDidLoad() {
     super.viewDidLoad()
 
-    view.backgroundColor = .clear
-    view.addSubview(backgroundView)
-    view.addSubview(contentNavigationController.view)
+    contentView.addSubview(contentNavigationController.view)
 
     updateLayoutConstraints()
   }
@@ -60,10 +52,6 @@ class SendTabToSelfController: UIViewController {
   }
   
   private func updateLayoutConstraints() {
-    backgroundView.snp.makeConstraints {
-      $0.edges.equalToSuperview()
-    }
-
     let preferredSize = sendTabContentController.view.systemLayoutSizeFitting(
       CGSize(width: view.bounds.size.width, height: view.frame.height),
       withHorizontalFittingPriority: .required,
@@ -74,7 +62,7 @@ class SendTabToSelfController: UIViewController {
     
     contentNavigationController.view.snp.makeConstraints {
       if traitCollection.horizontalSizeClass == .compact && traitCollection.verticalSizeClass == .regular {
-        $0.leading.trailing.equalTo(view.safeAreaLayoutGuide).inset(UX.contentInset)
+        $0.leading.trailing.equalTo(contentView.safeAreaLayoutGuide).inset(UX.contentInset)
       } else {
         $0.width.equalToSuperview().multipliedBy(0.75)
       }
